LPS-173429 Improve the package naming
